e3: ots_getContractCreator (#854)

+++ kv/kv_interface.go
@@ -460,7 +460,14 @@ type TemporalTx interface {
 	DomainGet(name Domain, k, k2 []byte, ts uint64) (v []byte, ok bool, err error)
 	HistoryGet(name History, k []byte, ts uint64) (v []byte, ok bool, err error)
 
-	IndexRange(name InvertedIdx, k []byte, fromTs, toTs uint64, asc order.By, limit int) (timestamps iter.U64, err error)
+	// IndexRange - return iterator over range of inverted index for given key `k`
+	// Asc semantic:  [from, to) AND from > to
+	// Desc semantic: [from, to) AND from < to
+	// Limit -1 means Unlimited
+	// from -1, to -1 means unbounded (StartOfTable, EndOfTable)
+	// Example: IndexRange("IndexName", 10, 5, order.Desc, -1)
+	// Example: IndexRange("IndexName", -1, -1, order.Asc, 10)
+	IndexRange(name InvertedIdx, k []byte, fromTs, toTs int, asc order.By, limit int) (timestamps iter.U64, err error)
 }
 
 type TemporalRwDB interface {
